% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_info.R
\docType{data}
\name{vz_data_2023}
\alias{vz_data_2023}
\title{Metadata for vehicle fatalities in Boston, NYC, Denver, Richmond, and Boulder ranging from years 2008-2023.}
\format{
A data frame with 36 rows and 5 columns:
\describe{
\item{year}{year when fatalities occurred}
\item{total_fatal}{total number of fatalities in a given year}
\item{city}{city where fatalities took place}
\item{pop_size}{number of population in a given city}
\item{fatal_perc}{percentage of fatalities per capita}
...
}

A data frame with 36 rows and 5 columns:
\describe{
\item{year}{Year that fatalities occured}
\item{total_fatal}{Number of documented fatalities resulting from a car crash per year as per each cities respective VisionZero data set. Note that some cities have a specific fatalities data set and other cities had a more general one that needed to be filtered for fatalities}
\item{city}{VisionZero city where the vehicular fatalities took place. Note that this is NOT an exhausive list of ALL VisionZero cities involved in the program.}
\item{pop_size}{Population size for city as of 2022}
\item{fatal_perc}{Percent of each citiy's population involved in vehicular fatalities per year calcuated with VisionZero data and Census population data}
...
}
}
\source{
<https://www.virginiaroads.org/datasets/VDOT::crashdata-details-3.csv,
https://opendata.arcgis.com/datasets/4918642e10c84c7fbb8d8590c5b9b216_0.csv,
https://data.cityofnewyork.us/api/views/h9gi-nx95/rows.csv?date=20231129&accessType=DOWNLOAD,
https://www.denvergov.org/media/gis/DataCatalog/traffic_accidents/csv/traffic_accidents.csv,
https://data.boston.gov/dataset/d326a4e3-75f2-42ac-9b32-e2920566d04c/resource/92f18923-d4ec-4c17-9405-4e0da63e1d6c/download/tmp1u21fr8k.csv,
*** url for populations?>

\url{https://data.boston.gov/dataset/d326a4e3-75f2-42ac-9b32-e2920566d04c/resource/92f18923-d4ec-4c17-9405-4e0da63e1d6c/download/tmp1u21fr8k.csv}
\url{https://data.cityofnewyork.us/api/views/h9gi-nx95/rows.csv?date=20231129&accessType=DOWNLOAD}
\url{https://www.denvergov.org/media/gis/DataCatalog/traffic_accidents/csv/traffic_accidents.csv}
\url{https://opendata.arcgis.com/datasets/4918642e10c84c7fbb8d8590c5b9b216_0.csv}
\url{https://www.virginiaroads.org/datasets/VDOT::crashdata-details-3.csv}
\url{https://www.census.gov/quickfacts/fact/csv/bostoncitymassachusetts,richmondcityvirginia,denvercountycolorado,bouldercitycolorado,newyorkcitynewyork/PST045222}
}
\usage{
vz_data_2023

vz_data_2023
}
\description{
Metadata for vehicle fatalities in Boston, NYC, Denver, Richmond, and Boulder ranging from years 2008-2023.

Data for VisionZero crash fatalities and respective city population data
}
\keyword{datasets}
