% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{vz_graph}
\alias{vz_graph}
\title{Creates point graph of vehicular fatalities normalized to the population size for user-given VisionZero cities}
\usage{
vz_graph(cities)
}
\arguments{
\item{character}{A character vector of up to 5 elements containing city names from five different VisionZero cities: Boston, NYC, Boulder, Denver, and Richmond.}
}
\value{
A point graph in the \code{plots} window displaying the percentage of vehicular fatalities based on population size across the years 2008-2023 (year range dependent on the available data for the user-given cities).
}
\description{
Given 1 to 5 VisionZero cities (Boston, NYC, Boulder, Denver or Richmond), this function creates a point graph of vehicular fatalities throughout the years (from 2008-2023) normalized to the population size of the city for the user-given cities. This function uses vehicular fatality and population data aggregated for these cities in the \code{vz_data_2023} data set
}
\examples{
# Graphing vehicular fatality data for just Richmond
vz_graph("Richmond")
# Graphing vehicular fatality data for Boston and Boulder
vz_graph(c("Boston", "Boulder"))

}
